+++ src/datastation/dataverse/dataverse_client.py
@@ -3,6 +3,7 @@ from datastation.dataverse.banner_api import BannerApi
 from datastation.dataverse.builtin_users import BuiltInUsersApi
 from datastation.dataverse.file_api import FileApi
 from datastation.dataverse.dataset_api import DatasetApi
+from datastation.dataverse.dataverse_api import DataverseApi
 
 
 class DataverseClient:
@@ -20,6 +21,9 @@ class DataverseClient:
 
     def dataset(self, pid):
         return DatasetApi(pid, self.server_url, self.api_token, self.unblock_key, self.safety_latch)
+
+    def dataverse(self):
+        return DataverseApi(self.server_url, self.api_token)
 
     def file(self, id):
         return FileApi(id, self.server_url, self.api_token, self.unblock_key, self.safety_latch)
